refactor: data 계층 패키지 위치 변경
