--- android/app/src/main/java/com/daedan/festabook/data/datasource/remote/NoticeDataSource.kt
@@ -1,7 +1,0 @@
-package com.daedan.festabook.data.datasource.remote
-
-import com.daedan.festabook.data.model.response.NoticeListResponse
-
-interface NoticeDataSource {
-    suspend fun fetchNotices(): ApiResult<NoticeListResponse>
-}
