+++ android/app/src/main/java/com/daedan/festabook/data/datasource/remote/notice/NoticeDataSourceImpl.kt
@@ -1,6 +1,7 @@
-package com.daedan.festabook.data.datasource.remote
+package com.daedan.festabook.data.datasource.remote.notice
 
-import com.daedan.festabook.data.model.response.NoticeListResponse
+import com.daedan.festabook.data.datasource.remote.ApiResult
+import com.daedan.festabook.data.model.response.notice.NoticeListResponse
 import com.daedan.festabook.data.service.NoticeService
 
 class NoticeDataSourceImpl(
