+++ android/app/src/main/java/com/daedan/festabook/data/datasource/remote/organization/OrganizationBookmarkDataSourceImpl.kt
@@ -1,7 +1,8 @@
-package com.daedan.festabook.data.datasource.remote
+package com.daedan.festabook.data.datasource.remote.organization
 
+import com.daedan.festabook.data.datasource.remote.ApiResult
 import com.daedan.festabook.data.model.request.OrganizationBookmarkRequest
-import com.daedan.festabook.data.model.response.OrganizationBookmarkResponse
+import com.daedan.festabook.data.model.response.organization.OrganizationBookmarkResponse
 import com.daedan.festabook.data.service.OrganizationBookmarkService
 
 class OrganizationBookmarkDataSourceImpl(
