+++ android/app/src/main/java/com/daedan/festabook/data/datasource/remote/place/PlaceDataSourceImpl.kt
@@ -1,10 +1,10 @@
 package com.daedan.festabook.data.datasource.remote.place
 
 import com.daedan.festabook.data.datasource.remote.ApiResult
-import com.daedan.festabook.data.model.response.OrganizationGeographyResponse
-import com.daedan.festabook.data.model.response.PlaceDetailResponse
-import com.daedan.festabook.data.model.response.PlaceGeographyResponse
-import com.daedan.festabook.data.model.response.PlaceResponse
+import com.daedan.festabook.data.model.response.organization.OrganizationGeographyResponse
+import com.daedan.festabook.data.model.response.place.PlaceDetailResponse
+import com.daedan.festabook.data.model.response.place.PlaceGeographyResponse
+import com.daedan.festabook.data.model.response.place.PlaceResponse
 import com.daedan.festabook.data.service.OrganizationService
 import com.daedan.festabook.data.service.PlaceService
