+++ android/app/src/main/java/com/daedan/festabook/data/datasource/remote/schedule/ScheduleDataSourceImpl.kt
@@ -1,8 +1,8 @@
 package com.daedan.festabook.data.datasource.remote.schedule
 
 import com.daedan.festabook.data.datasource.remote.ApiResult
-import com.daedan.festabook.data.model.response.ScheduleDateResponse
-import com.daedan.festabook.data.model.response.ScheduleEventResponse
+import com.daedan.festabook.data.model.response.schedule.ScheduleDateResponse
+import com.daedan.festabook.data.model.response.schedule.ScheduleEventResponse
 import com.daedan.festabook.data.service.ScheduleService
 
 class ScheduleDataSourceImpl(
