+++ android/app/src/main/java/com/daedan/festabook/data/model/response/place/PlaceDetailResponse.kt
@@ -1,4 +1,4 @@
-package com.daedan.festabook.data.model.response
+package com.daedan.festabook.data.model.response.place
 
 import com.daedan.festabook.domain.model.Notice
 import com.daedan.festabook.domain.model.Place
