+++ android/app/src/main/java/com/daedan/festabook/data/repository/BookmarkRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.daedan.festabook.data.repository
 
-import com.daedan.festabook.data.datasource.remote.OrganizationBookmarkDataSource
+import com.daedan.festabook.data.datasource.remote.organization.OrganizationBookmarkDataSource
 import com.daedan.festabook.data.util.toResult
 import com.daedan.festabook.domain.repository.BookmarkRepository
 
@@ -22,7 +22,8 @@ class BookmarkRepositoryImpl(
     }
 
     override suspend fun deleteOrganizationBookmark(organizationBookmarkId: Long): Result<Unit> {
-        val response = organizationBookmarkDataSource.deleteOrganizationBookmark(organizationBookmarkId)
+        val response =
+            organizationBookmarkDataSource.deleteOrganizationBookmark(organizationBookmarkId)
         val result = response.toResult()
 
         return result
