+++ android/app/src/main/java/com/daedan/festabook/data/service/NoticeService.kt
@@ -1,6 +1,6 @@
 package com.daedan.festabook.data.service
 
-import com.daedan.festabook.data.model.response.NoticeListResponse
+import com.daedan.festabook.data.model.response.notice.NoticeListResponse
 import retrofit2.Response
 import retrofit2.http.GET
